refactor: update game feature public api
