+++ src/features/game/index.ts
@@ -1,2 +1,5 @@
 export { default as RoundControls } from "./ui/controls/RoundControls";
 export { default as StageControls } from "./ui/controls/StageControls";
+export { default as LevelsState } from "./model/LevelsState";
+export { default as RoundState, type Stage } from "./model/RoundState";
+export { StageStatus } from "./model/StageStatus";
