+++ src/widgets/fields/ui/GameField.ts
@@ -1,7 +1,7 @@
 import { Component } from "../../../shared/ui";
 import AssembleRow from "../../../features/game/fields/AssembleRow";
 
-import RoundState from "../../../features/game/model/RoundState";
+import { RoundState } from "../../../features/game";
 import { HintSettings } from "../../../features/hints";
 
 import { Observer, Publisher } from "../../../entities/state";
